updated the new project script
to include non random database credentials

+++ new_project.bash
@@ -1,19 +1,14 @@
 #!/bin/bash
 # Build Project Directories
-# v4.3
-# Last Updated: Dec 2, 2012
+# v5.0
+# Last Updated: Dec 19, 2012
 # Documentation: 
 # http://www.nickyeoman.com/blog/system-administration/18-project-directory-setup
 
-#Use like this: bash new_project.bash domainName.com
-
-# CHANGELOG for 4.0
-# Modified to work with plesk on CentOS
-# dropped the use of symlinks for increased compatibility 
-# plesk has one click installs, so we are only installing Joomla now
-# We still want to run git so we are running from git/public instead of default httpdocs
+#Use like this: bash new_project.bash domainName.com dbname dbuser dbpass
 
 # REQUIREMENTS
+# Joomla 2.5.8
 # Ubuntu/debian:
 # sudo apt-get install php-cli
 # You also need an internet connection
@@ -23,7 +18,7 @@
 ##
 	projectDir=/git #full path to install directory
 	salt=$RANDOM #Change this to something static for recoverable passwords
-	db_prefix="admin_"
+	sinstall=http://joomlacode.org/gf/download/frsrelease/17715/77262/Joomla_2.5.8-Stable-Full_Package.zip
 	
 	#Project Domain
 	if [ -z "$1" ]; then
@@ -32,33 +27,56 @@
 	else
 	  domain=$1
 	fi
-
-##
-# Create Directories
-##
+	
+	##
+	# Create Directories
+	##
 	cd $projectDir
 	#TODO: check for domain then error if exists
 	mkdir $domain
 	cd $projectDir/$domain
 	mkdir scripts sql apache
+	
+	#db name
+	if [ -z "$2" ]; then
+		dbname=`echo $domain | sed 's/\(.*\)\..*/\1/'`
+	else
+		dbname=$2
+	fi
+	
+	#db user
+	if [ -z "$3" ]; then
+		dbuser=`echo $dbname`
+	else
+		dbuser=$3
+	fi
+	
+	#dbpass
+	if [ -z "$4" ]; then
+		cd sql
+		wget https://raw.github.com/nickyeoman/NYScripts/master/sha1.php
+		dbpass=`php sha1.php $domain $salt $salt` 
+		rm sha1.php
+		cd $projectDir
+	else
+		dbuser=$4
+	fi
 
 ##
 # Grab Nick Yeoman's scripts
 ##
 	cd scripts
 	wget https://raw.github.com/nickyeoman/NYScripts/master/database.bash
-	wget https://raw.github.com/nickyeoman/NYScripts/master/config.sh
+	wget https://raw.github.com/nickyeoman/NYScripts/master/config.bash
+	cat <<xFiledumpx > dump_db.bash
+./database.bash d $dbname $dbuser $dbpass localhost
+xFiledumpx
+	cat <<xFileupdatex > update_db.bash
+./database.bash u $dbname $dbuser $dbpass localhost
+xFileupdatex
+
 	cd $projectDir/$domain
 
-##
-#Create DB script
-##
-	cd sql
-	wget https://raw.github.com/nickyeoman/NYScripts/master/sha1.php
-	dbpass=`php sha1.php $domain $salt $salt` 
-	dbname=`echo $db_prefix$domain | sed 's/\(.*\)\..*/\1/'`
-	dbuser=`echo $dbname`
-	rm sha1.php
 	
 #--------Begin here document-----------#
 cat <<xFileconfigsqlx > config.sql
@@ -79,11 +97,12 @@ xFileconfigsqlx
 # Create Apache config file (config.sh uses this)
 ##
 	cd $projectDir/$domain/apache
+	$project=`echo $domain | sed 's/\(.*\)\..*/\1/'`
 #--------Begin here document-----------#	
 cat <<xFileconfigshx > $domain
 <VirtualHost *:80>
 	ServerName $domain
-	ServerAlias www.$domain *.$domain
+	ServerAlias www.$domain *.$domain $project.ny $project.fb
 	DocumentRoot $projectDir/$domain/public/
 	ServerAdmin webmaster@$domain
 	
@@ -106,11 +125,11 @@ xFileconfigshx
 	mkdir $projectDir/$domain/public
 	cd $projectDir/$domain/public
 	
-	wget http://joomlacode.org/gf/download/frsrelease/17715/77262/Joomla_2.5.8-Stable-Full_Package.zip
-	unzip -e Joomla_2.5.8-Stable-Full_Package.zip
+	wget $sinstall
+	unzip -e *.zip
 	rm -rf *.zip
 	cd $projectDir/$domain
-
+	
 ##
 # All Done
 ##
@@ -120,16 +139,13 @@ cat <<xtalkToMex
 ****************************************
 Installation Finished
 Your domain ($domain) is setup
-Directory = $projectDir/$domain
-Database name = $dbname
-Database user = $dbuser
-Databse password = $dbpass
-(you can refer to config.sql for this info)
 
 Notes: 
-* Fix permissions
-* git init
-* You can now run config.sh on your debian based dev server (as root)
+* Run config.bash to setup database
+* Update .htaccess (php, non-www, redirects)
+* Go to the web to run the web installer
+* Install frosty apps
+* git init and push
 ****************************************
 xtalkToMex
